requirements.txt
telethon
psycopg2
asyncpg
aiohttp
beautifulsoup4
python-telegram-bot
